+++ lib/oxidized/source/http.rb
@@ -22,7 +22,7 @@ module Oxidized
       uri = URI.parse(@cfg.url)
       data = JSON.parse(read_http(uri, node_want))
       node_data = data
-      node_data = string_navigate(data, @cfg.hosts_location) if @cfg.hosts_location?
+      node_data = string_navigate_object(data, @cfg.hosts_location) if @cfg.hosts_location?
       node_data = pagination(data, node_want) if @cfg.pagination?
 
       # at this point we have all the nodes; pagination or not
@@ -32,7 +32,7 @@ module Oxidized
         # map node parameters
         keys = {}
         @cfg.map.each do |key, want_position|
-          keys[key.to_sym] = node_var_interpolate string_navigate(node, want_position)
+          keys[key.to_sym] = node_var_interpolate string_navigate_object(node, want_position)
         end
         keys[:model] = map_model keys[:model] if keys.has_key? :model
         keys[:group] = map_group keys[:group] if keys.has_key? :group
@@ -40,7 +40,7 @@ module Oxidized
         # map node specific vars
         vars = {}
         @cfg.vars_map.each do |key, want_position|
-          vars[key.to_sym] = node_var_interpolate string_navigate(node, want_position)
+          vars[key.to_sym] = node_var_interpolate string_navigate_object(node, want_position)
         end
         keys[:vars] = vars unless vars.empty?
 
@@ -51,29 +51,18 @@ module Oxidized
 
     private
 
-    def string_navigate(object, wants)
-      wants = wants.split(".").map do |want|
-        head, match, _tail = want.partition(/\[\d+\]/)
-        match.empty? ? head : [head, match[1..-2].to_i]
-      end
-      wants.flatten.each do |want|
-        object = object[want] if object.respond_to? :each
-      end
-      object
-    end
-
     def pagination(data, node_want)
       node_data = []
       raise Oxidized::OxidizedError, "if using pagination, 'pagination_key_name' setting must be set" unless @cfg.pagination_key_name?
 
       next_key = @cfg.pagination_key_name
       loop do
-        node_data += string_navigate(data, @cfg.hosts_location) if @cfg.hosts_location?
+        node_data += string_navigate_object(data, @cfg.hosts_location) if @cfg.hosts_location?
         break if data[next_key].nil?
 
         new_uri = URI.parse(data[next_key]) if data.has_key?(next_key)
         data = JSON.parse(read_http(new_uri, node_want))
-        node_data += string_navigate(data, @cfg.hosts_location) if @cfg.hosts_location?
+        node_data += string_navigate_object(data, @cfg.hosts_location) if @cfg.hosts_location?
       end
       node_data
     end
