+++ lib/oxidized/source/jsonfile.rb
@@ -1,7 +1,5 @@
 module Oxidized
-  require "csv"
   class JSONFile < Source
-
     def initialize
       @cfg = Oxidized.config.source.jsonfile
       super
@@ -21,17 +19,17 @@ module Oxidized
 
     require "json"
 
-    def load(node_want = nil)
+    def load(*)
       nodes = []
       data = JSON.parse(open_file.read)
-      data = string_navigate(data, @cfg.hosts_location) if @cfg.hosts_location?
+      data = string_navigate_object(data, @cfg.hosts_location) if @cfg.hosts_location?
       data.each do |node|
         next if node.empty?
 
         # map node parameters
         keys = {}
         @cfg.map.each do |key, want_position|
-          keys[key.to_sym] = node_var_interpolate string_navigate(node, want_position)
+          keys[key.to_sym] = node_var_interpolate string_navigate_object(node, want_position)
         end
         keys[:model] = map_model keys[:model] if keys.has_key? :model
         keys[:group] = map_group keys[:group] if keys.has_key? :group
@@ -39,36 +37,13 @@ module Oxidized
         # map node specific vars
         vars = {}
         @cfg.vars_map.each do |key, want_position|
-          vars[key.to_sym] = node_var_interpolate string_navigate(node, want_position)
+          vars[key.to_sym] = node_var_interpolate string_navigate_object(node, want_position)
         end
         keys[:vars] = vars unless vars.empty?
 
         nodes << keys
       end
       nodes
-    end
-
-    private
-
-    def string_navigate(object, wants)
-      wants = wants.split(".").map do |want|
-        head, match, _tail = want.partition(/\[\d+\]/)
-        match.empty? ? head : [head, match[1..-2].to_i]
-      end
-      wants.flatten.each do |want|
-        object = object[want] if object.respond_to? :each
-      end
-      object
-    end
-
-    def open_file
-      file = File.expand_path(@cfg.file)
-      if @cfg.gpg?
-        crypto = GPGME::Crypto.new password: @cfg.gpg_password
-        crypto.decrypt(File.open(file)).to_s
-      else
-        File.open(file)
-      end
     end
   end
 end
