+++ lib/oxidized/source/source.rb
@@ -23,5 +23,28 @@ module Oxidized
       else var
       end
     end
+
+    private
+
+    def open_file
+      file = File.expand_path(@cfg.file)
+      if @cfg.gpg?
+        crypto = GPGME::Crypto.new password: @cfg.gpg_password
+        crypto.decrypt(File.open(file)).to_s
+      else
+        File.open(file)
+      end
+    end
+
+    def string_navigate_object(object, wants)
+      wants = wants.split(".").map do |want|
+        head, match, _tail = want.partition(/\[\d+\]/)
+        match.empty? ? head : [head, match[1..-2].to_i]
+      end
+      wants.flatten.each do |want|
+        object = object[want] if object.respond_to? :each
+      end
+      object
+    end
   end
 end
